compiler/pyproject.toml
[tool.poetry]
name = "compiler"
version = "0.1.0"
description = ""
authors = ["Adam Zahradník <adam@zahradnik.xyz>"]

[tool.poetry.dependencies]
python = "^3.11"
podman = "^4.7.0"
pyyaml = "^6.0.1"
celery = "^5.3.5"
redis = "^5.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
